On the way to memory-based RL training interface.
Added the snapshot/restore methods to the core gym class Env interface
so that we can record state of the environment & agent at any time step
and restore the system according to the snapshot. Here we assume that
the system is Markovian -- the evolution from time step t on wards only
depends on the system state at time step t but not on the history.

[FIX ME] Added deepcopy() method to the agent class Car() but does not
work right now because the Box2d objects cannot be deep-copied.

[FIX ME] In the CarRacing(gym.Env) class, I tried to implement the
_snapshot() and _restore() functions that required in the new interface.
It should be easily fixed once class Car() can be deep copied.